ingestion/pyproject.toml
[project]
name = "ingestion"
version = "0.1.0"
description = ""
authors = [
    {name = "dengke-li",email = "linuslukia@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "psycopg (>=3.3.2,<4.0.0)",
    "psycopg-binary (>=3.3.2,<4.0.0)",
    "feedparser (>=6.0.12,<7.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "ingest_etl"}]

# Add this section for dev tools
[tool.poetry.group.dev.dependencies]
pytest = ">=9.0.2,<10.0.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
